WIP added logger, model package moved to pkg
