+++ cmd/client/main.go
@@ -1,7 +1,22 @@
 package main
 
-import "fmt"
+import (
+	"context"
+	"fmt"
+	"log"
+
+	"github.com/dmad1989/gophKeeper/pkg/logger"
+	"github.com/dmad1989/gophKeeper/pkg/model/consts"
+)
 
 func main() {
+	zlog, err := logger.NewLogger()
+	if err != nil {
+		log.Fatal(err)
+	}
+	zlog = zlog.Named("client")
+	ctx := context.WithValue(context.Background(), consts.LoggerCtxKey, zlog)
+	defer zlog.Sync()
+
 	fmt.Println("client is working!")
 }
