+++ internal/server/repository/repository.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/dmad1989/gophKeeper/internal/model/consts"
 	"github.com/dmad1989/gophKeeper/internal/server/repository/db"
+	"github.com/dmad1989/gophKeeper/pkg/model/consts"
 	"github.com/jackc/pgx/v5"
 	"go.uber.org/zap"
 )
@@ -17,6 +17,8 @@ type repo struct {
 }
 
 func New(ctx context.Context, c string) (*repo, error) {
+	log := ctx.Value(consts.LoggerCtxKey).(*zap.Logger).Named("repository")
+
 	pconf, err := pgx.ParseConfig(c)
 	if err != nil {
 		return nil, fmt.Errorf("repository.new: ParseConfig: %w", err)
@@ -28,7 +30,7 @@ func New(ctx context.Context, c string) (*repo, error) {
 	}
 
 	return &repo{
-		ctx.Value(consts.LoggerCtxKey).(*zap.Logger),
+		log,
 		db.New(conn),
 		conn,
 	}, nil
