+++ pkg/logger/logger.go
@@ -1,0 +1,18 @@
+package logger
+
+import (
+	"fmt"
+
+	"go.uber.org/zap"
+)
+
+func NewLogger() (*zap.SugaredLogger, error) {
+	cfg := zap.NewProductionConfig()
+	cfg.DisableStacktrace = true
+	cfg.Level = zap.NewAtomicLevelAt(zap.DebugLevel)
+	logger, err := cfg.Build()
+	if err != nil {
+		return nil, fmt.Errorf("logger.NewLogger: cfg.Build: %w", err)
+	}
+	return logger.Sugar(), nil
+}
